requirements.txt
configparser==5.0.1
PySimpleGUI==4.32.1
python-crontab==2.5.1
python-dateutil==2.8.1
six==1.15.0
watchdog==1.0.2
